+++ varfish/static/js/state_machine.js
@@ -380,7 +380,7 @@ function handleEventStateWaitJobResults(eventType, event) {
               "order": [[ 1, "asc" ]],
               'aoColumnDefs': [
                 {
-                  'aTargets': [0,2,3,5,6,8,9,27,-1], /* column index */
+                  'aTargets': [0,2,3,5,6,8,9,28,-1], /* column index */
                   'bSortable': false, /* true or false */
                 },
               ]