+++ HISTORY.rst
@@ -12,6 +12,7 @@ End-User Summary
 - Added caching for pathogenicity scores api results.
 - Added column to the project wide filter results table that displays the number of affected cases per gene.
 - Enabled pathogenicity scoring for project-wide filtration.
+- Added LOEUF gnomAD constraint column to results table.
 
 Full Change List
 ================
@@ -21,6 +22,7 @@ Full Change List
   I.e. the cases (not samples) that have a variant in a gene are counted and reported.
 - Enabled pathogenicity scoring for project-wide filtration.
   This introduced a new table ``ProjectCasesSmallVariantQueryVariantScores`` to store the scoring results for a query.
+- Added LOEUF gnomAD constraint column to results table.
 
 -------
 v0.17.6